test(integration): introduce e2e folder + dotenv loading
- Move batchTransactions.test.js to test/integration/e2e/batch-transactions/
  to mirror the layout of the abstractionkit-examples repo.
- Add test/integration/_runnable.cjs helper exporting runnable/unrunnable
  chain lists and nodeUrl/bundlerUrl shortcuts, so each e2e test stops
  repeating the status-file boilerplate.
- Re-add dotenv as a devDep; globalSetup.cjs and _runnable.cjs now load
  .env at startup, so ETHEREUM_RPC / OPTIMISM_RPC / BASE_RPC overrides
  can live in a local .env file instead of being passed inline.
- Add .env.example documenting the optional RPC override variables.
