+++ .env.example
@@ -1,0 +1,8 @@
+# All variables are optional — `yarn test:integration` runs with sensible defaults.
+
+# ---------- Fork RPCs ----------
+# Override the default public RPCs used by anvil to fork each chain.
+# Defaults: ethereum-rpc.publicnode.com, optimism-rpc.publicnode.com, mainnet.base.org
+ETHEREUM_RPC=
+OPTIMISM_RPC=
+BASE_RPC=
