+++ package.json
@@ -63,6 +63,7 @@
 	},
 	"devDependencies": {
 		"@biomejs/biome": "^2.4.12",
+		"dotenv": "^16.4.5",
 		"jest": "^29.7.0",
 		"tsdown": "^0.21.6",
 		"typescript": "^5.1.6"
